+++ src/Modules/Admin/Widgets/Forms/Traits/ParentIdFieldTrait.php
@@ -16,7 +16,6 @@ trait ParentIdFieldTrait
         }
 
         return $this->createParentIdSelectField()
-            ->attributes($this->getParentIdAttributes())
             ->property('parent_id')
             ->prompt();
     }
@@ -24,12 +23,4 @@ trait ParentIdFieldTrait
     abstract protected function createParentIdSelectField(): SelectField;
 
     abstract protected function hasParentIdField(): bool;
-
-    /**
-     * @return array<string, mixed>
-     */
-    protected function getParentIdAttributes(): array
-    {
-        return [];
-    }
 }